Updates for work within linux environment

+++ files/setup/setup.sh
@@ -10,19 +10,28 @@ usage() {
 
 [[ $EUID -ne 0 ]] && usage "Need to run as root."
 
+echo "Load useradd settings"
+. /etc/default/useradd
+
 echo "Removing items from chkconfig"
-chkconfig --del duke
-chkconfig --del nuke
+/sbin/chkconfig --del duke 2>/dev/null || echo "duke not installed"
+/sbin/chkconfig --del nuke 2>/dev/null || echo "nuke not installed"
 
 echo "Removing old files"
-rm -rf /etc/opus
-rm -rf /usr/share/java/opus
-rm -rf /usr/share/javadoc/opus
-rm -rf /usr/log/opus
+[[ -d /etc/opus ]] && rm -rf /etc/opus
+[[ -d /usr/share/java/opus ]] && rm -rf /usr/share/java/opus
+[[ -d /usr/share/javadoc/opus ]] && rm -rf /usr/share/javadoc/opus
+[[ -d /var/log/opus ]] && rm -rf /var/log/opus
 
 echo "Creating user"
-getent group opus || groupadd -r opus
-getent passwd opus || useradd -r -d /home/opus -g opus -s /bin/bash opus
+/usr/bin/getent group opus || /usr/sbin/groupadd -r opus
+/usr/bin/getent passwd opus || /usr/sbin/useradd -r -d /home/opus -g opus -s /bin/bash opus
+
+if [ ! -d /home/opus ]; then
+  echo "useradd did not create home directory for opus user"
+  mkdir -p /home/opus
+  /bin/chown opus:opus -R /home/opus
+fi
 
 echo "Creating directories"
 mkdir -p {/etc/opus,/usr/share/java/opus,/usr/share/javadoc/opus,/var/log/opus/nuke,/var/log/opus/duke}
@@ -40,12 +49,12 @@ ln -sf /usr/share/java/opus/duke*.jar /usr/share/java/opus/duke.jar
 ln -sf /usr/share/java/opus/nuke*.jar /usr/share/java/opus/nuke.jar
 
 echo "Adding items to chkconfig"
-chkconfig --add nuke
-chkconfig --add duke
+/sbin/chkconfig --add nuke
+/sbin/chkconfig --add duke
 
 echo "Setting up folder permissions"
-chown opus:opus -R /etc/opus
-chown opus:opus -R /usr/share/java/opus
-chown opus:opus -R /usr/share/javadoc/opus
-chown opus:opus -R /var/log/opus
+/bin/chown opus:opus -R /etc/opus
+/bin/chown opus:opus -R /usr/share/java/opus
+/bin/chown opus:opus -R /usr/share/javadoc/opus
+/bin/chown opus:opus -R /var/log/opus
